+++ e2e/interactions.test.ts
@@ -16,7 +16,7 @@ import {
   waitForChannelPost,
   waitForFollowUp,
 } from './setup';
-import { ReminderDurableObject } from '../src/reminder.js';
+import { ReminderDurableObject } from '../src/commands/reminder.js';
 
 describe('Discord Worker', () => {
   it('responds to Discord Ping interaction', async () => {
