+++ src/commands/wordOfDaySchedule.ts
@@ -3,8 +3,8 @@ import {
   fetchWordOfDayEntry,
   formatWordOfDayMessage,
   WORD_OF_DAY_DEFAULT_FEED_URL,
-} from './wordOfDay.js';
-import { sendDiscordMessage } from './skills/sendDiscordMessage.js';
+} from '../wordOfDay.js';
+import { sendDiscordMessage } from '../skills/sendDiscordMessage.js';
 
 const WORD_OF_DAY_POST_HOUR = 7;
 const WORD_OF_DAY_POST_MINUTE = 30;
