+++ src/index.ts
@@ -1,2 +1,2 @@
 export { default } from './app.js';
-export { ReminderDurableObject } from './reminder.js';
+export { ReminderDurableObject } from './commands/reminder.js';
