+++ src/integrations/github/issueClient.ts
@@ -1,40 +1,10 @@
-import type {
-  FollowUpExecutionContext,
-  FollowUpExecutionResult,
-  FollowUpTask,
-} from './core/index.js';
-import { extractModalFields, type ModalComponentRows } from './modal.js';
-
-export const ISSUE_COMMAND_NAME = 'issue';
-export const ISSUE_MODAL_ID = 'issue_modal';
-export const ISSUE_TITLE_INPUT_ID = 'issue_title';
-export const ISSUE_BODY_INPUT_ID = 'issue_body';
-
 const GITHUB_API_VERSION = '2022-11-28';
 const GITHUB_USER_AGENT = 'discord-worker (+https://github.com/orryjarvis/discord-worker)';
-const GITHUB_ISSUE_FAILURE_MESSAGE = 'Could not create GitHub issue right now. Try again in a moment.';
 
 type GitHubIssueRepository = {
   owner: string;
   repo: string;
 };
-
-export interface IssueRuntimeEnv {
-  GITHUB_APP_ID: string;
-  GITHUB_APP_INSTALLATION_ID: string;
-  GITHUB_APP_PRIVATE_KEY: string;
-  GITHUB_ISSUE_REPOSITORY: string;
-  GITHUB_API_BASE_URL?: string;
-}
-
-export type IssueModalParseResult =
-  | { kind: 'unknown-modal' }
-  | { kind: 'missing-fields' }
-  | {
-    kind: 'parsed';
-    commandName: typeof ISSUE_COMMAND_NAME;
-    fields: Record<string, string>;
-  };
 
 type GitHubInstallationTokenResponse = {
   token?: string;
@@ -44,6 +14,14 @@ type GitHubCreatedIssueResponse = {
   html_url?: string;
   number?: number;
 };
+
+export interface GitHubIssueClientEnv {
+  GITHUB_APP_ID: string;
+  GITHUB_APP_INSTALLATION_ID: string;
+  GITHUB_APP_PRIVATE_KEY: string;
+  GITHUB_ISSUE_REPOSITORY: string;
+  GITHUB_API_BASE_URL?: string;
+}
 
 function parseIssueRepository(repository: string): GitHubIssueRepository {
   const [owner, repo, ...rest] = repository.split('/').filter(Boolean);
@@ -97,14 +75,6 @@ function decodePemToArrayBuffer(pem: string): ArrayBuffer {
   return buffer;
 }
 
-function describeError(error: unknown): string {
-  if (error instanceof Error) {
-    return `${error.name}: ${error.message}`;
-  }
-
-  return String(error);
-}
-
 async function readResponseText(response: Response): Promise<string> {
   const text = (await response.text()).trim();
   if (!text) {
@@ -123,32 +93,7 @@ async function readResponseText(response: Response): Promise<string> {
   return text.length > 200 ? `${text.slice(0, 200)}...` : text;
 }
 
-export function parseIssueModalSubmit(data: {
-  customId: string;
-  components?: ModalComponentRows;
-}): IssueModalParseResult {
-  if (data.customId !== ISSUE_MODAL_ID) {
-    return { kind: 'unknown-modal' };
-  }
-
-  const fields = extractModalFields(data.components);
-  const title = fields[ISSUE_TITLE_INPUT_ID]?.trim();
-  const body = fields[ISSUE_BODY_INPUT_ID]?.trim();
-  if (!title || !body) {
-    return { kind: 'missing-fields' };
-  }
-
-  return {
-    kind: 'parsed',
-    commandName: ISSUE_COMMAND_NAME,
-    fields: {
-      [ISSUE_TITLE_INPUT_ID]: title,
-      [ISSUE_BODY_INPUT_ID]: body,
-    },
-  };
-}
-
-async function createGitHubAppJwt(env: IssueRuntimeEnv): Promise<string> {
+async function createGitHubAppJwt(env: GitHubIssueClientEnv): Promise<string> {
   const now = Math.floor(Date.now() / 1000);
   const header = encodeBase64UrlJson({ alg: 'RS256', typ: 'JWT' });
   const payload = encodeBase64UrlJson({
@@ -182,7 +127,7 @@ async function createGitHubAppJwt(env: IssueRuntimeEnv): Promise<string> {
   return `${signingInput}.${toBase64Url(new Uint8Array(signature))}`;
 }
 
-async function createInstallationAccessToken(env: IssueRuntimeEnv, apiBaseUrl: string): Promise<string> {
+async function createInstallationAccessToken(env: GitHubIssueClientEnv, apiBaseUrl: string): Promise<string> {
   const jwt = await createGitHubAppJwt(env);
   const response = await fetch(`${apiBaseUrl}/app/installations/${env.GITHUB_APP_INSTALLATION_ID}/access_tokens`, {
     method: 'POST',
@@ -209,8 +154,8 @@ async function createInstallationAccessToken(env: IssueRuntimeEnv, apiBaseUrl: s
   return parsed.token;
 }
 
-async function createGitHubIssue(
-  env: IssueRuntimeEnv,
+export async function createGitHubIssue(
+  env: GitHubIssueClientEnv,
   title: string,
   body: string,
 ): Promise<{ htmlUrl: string; number: number }> {
@@ -248,40 +193,4 @@ async function createGitHubIssue(
     htmlUrl: parsed.html_url,
     number: parsed.number,
   };
-}
-
-export async function executeIssueFollowUp(
-  task: FollowUpTask,
-  env: IssueRuntimeEnv,
-  context: FollowUpExecutionContext,
-): Promise<FollowUpExecutionResult> {
-  const title = typeof task.payload.title === 'string' ? task.payload.title.trim() : '';
-  const body = typeof task.payload.body === 'string' ? task.payload.body.trim() : '';
-  if (!title || !body) {
-    console.warn('Issue follow-up payload missing title or body', {
-      messageId: context.messageId,
-      commandName: task.commandName,
-    });
-    return {
-      content: GITHUB_ISSUE_FAILURE_MESSAGE,
-    };
-  }
-
-  try {
-    const issue = await createGitHubIssue(env, title, body);
-    return {
-      content: `Created GitHub issue #${issue.number}: ${issue.htmlUrl}`,
-    };
-  } catch (error) {
-    console.error('GitHub issue creation failed', {
-      messageId: context.messageId,
-      commandName: task.commandName,
-      titleLength: title.length,
-      bodyLength: body.length,
-      error: describeError(error),
-    });
-    return {
-      content: GITHUB_ISSUE_FAILURE_MESSAGE,
-    };
-  }
 }
